collector: new module
Collectors are modules gathering information, mainly collecting events
and/or appropriate data; they are at the core of the tool. This adds an
initial implementation to allow registering collectors, using a Trait to
describe their common behaviour as well as defining a way to manipulate
a group of collectors (aka. the public API).

Please look at the patch itself for more insights about the
implementation (as it is documented in lengths there already).

Signed-off-by: Antoine Tenart <atenart@redhat.com>