core: kernel_symbols: refactor the kallsyms selection for tests
When running tests we need to use a known one so we can look up
addresses and symbols and be sure we'll find them and we'll find a fixed
value. To achieve this a different kallsyms file is used when running
tests.

The issue with the current implementation is tests using the module
can't work if the kernel_symbols tests didn't run first, because then
the kassyms singleton would not have been initialized with the test
data. To fix this move to a kallsyms file conditional definition
directly in the implementation. This is a build time choice.

This allows running tests which would have failed otherwise, such as:

  $ cargo test core::probe::kernel
  running 5 tests
  test core::probe::kernel::kprobe::tests::init_and_attach ... ignored
  test core::probe::kernel::raw_tracepoint::tests::init_and_attach ... ignored
  test core::probe::kernel::tests::reuse_map ... ok
  test core::probe::kernel::tests::add_probe ... ok
  test core::probe::kernel::tests::register_hooks ... ok

Signed-off-by: Antoine Tenart <atenart@redhat.com>