core: probe: kernel: inspect types and add per-type BPF helpers
This adds automatic inspection for some types (struct sk_buff, enum
skb_drop_reason, struct net_device, struct net) and configures the
probes so they can easily retrieve matching arguments, if any.

An helper is also added to check if a function has a given parameter as
an argument, given its type. This can be used by collectors as well to
dynamically inspect the kernel and/or configure their hooks.

The probe configuration part (config_map creation and use) is protected
by #[cfg(not(test))] so that it does not run when starting unit tests.
This should be OK for now, if this becomes a burden to maintain we can
move to a runtime solution.

Signed-off-by: Antoine Tenart <atenart@redhat.com>