core: probe: kernel: add an helper to retrieve the ksym of probes
Hooks have access to the current probed function ksym and can use this
information to know from where they are being called, one example being
retrieving a function-specific configuration. But this ksym isn't a 1:1
mapping with the actual symbol address, as there are special cases (for
instance raw tracepoints).

Export an helper so that modules registering hooks can get the ksym
value that will later be available in BPF.

Signed-off-by: Antoine Tenart <atenart@redhat.com>