proc: support USDT probes defined on shared libs
Programs might link statically or dynamically against third party
libraries that might contain USDT probes. This is currently painful for
users who have to figure out if the USDT they are looking for is defined
on a shared library and whether the target process linked statically or
dynamically against it.

To fix this, split the current process API in two:
- Binary: contains information about a binary which might be an
  executable or a library.
- Program: contains one executable Binary and zero or more library
  Binary objects while hiding where the USDT comes from.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>