scripts: generate split kernel header files
Typically, we use bpftool to generate headers from BTF files. However,
when we do that on a split file (i.e: module) all the types in vmlinux.h
are also included in the split file.

To avoid duplicating all the types in vmlinux.h for all kernel modules
that we will ever support, add a small script that deduplicates them and
generates split header files.

Use it to generate openvswitch header files:
    ./script/gen-gtf-headers.sh openvswitch

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>