user: bpf: Pre-fetch user arguments.
When we make a copy of the context it looses the CTX_PTR trait so it
cannot be sent down to helpers such as bpf_usdt_arg().

Also, the context itself cannot be sent through non-inline functions.

Pre fetch all the arguments and store them in a user_context that can be
passed down to hooks and used to retrieve arguments.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>