probe: kernel: add kretprobe support
Kretprobes work in a slightly differnt way compared with other kernel
probes.

Becuase function arguments are not available when the kretprobe
program is executed, we attach a small program to the kprobe of the same
symbol. This small kprobe stores the registers in a map that the
kretprobe then reads. Filtering can also be implemented by in the kprobe.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>