module: ovs: introduce queue_id to track upcalls
Tracking packets from an upcall queue event (kernel) to an upcall recv
event (userspace) is not trivial. With the current USDTs we cannot
really use other metadata (such as the socket id) to reliably correlate
upcalls being sent by the kernel and received by ovs-vswithchd.

Therefore, the approach taken here is to hash the first bytes of the
packet being sent and use that as "queue_id" which identifies the packet
in the upcall queue.

We could add the packet bytes to the event and perform the correlation
in post-processing time but that is very inefficient and increases the
number of missed events because of the added pressure on the event ring
buffer. Instead, this patch adds a simple header-only inline-able jhash
library taken from the linux kernel which makes it possible to perform
the hashing in the ebpf programs.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>