module: ovs: Add upcall to upcall enqueue tracking
One upcall event can be followed by one or more upcall queue events
because a packet an be split up into different segments before being
enqueued.

In order to track then we:
- Use a map to store the upcall context that can uniquely relate to a
  specific upcall. This information is stored by the upcall event.
- Retrieve this context and add it to the event on every enqueue event.
- Remove the context when the upcall ends.

As future improvement, we could use a standalone ebpf program (instead
of a hook) for the kretprobe that cleans the context.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>