module: ovs: command execute tracking
In order to track when an action execution comes from a command (i.e:
from userspace), add a kprobe and a kretprobe to ovs_packet_cmd_execute
to add and remove context simply indicating that the action corresponds
to a command. If so, the action execution hook adds another event
section containing the queue_id.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>