allow hooks to opt-in filters outcome
The outcome of some filters is not crucial for some modules, like OvS.
This patch will add the chance for hooks to ignore filtering outcome.

Given that filters are no longer causing ebpf programs to return,
events can now be discarded altogether if no hooks consumes any part
of it.
Thanks to Adrian for pointing that out and to both Adrian and Antoine
for brainstorming about it.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>