cli: move subcommand logic to its module
Add a trait called SubCommandRunner that can be implemented by simple
functions or closures. The SubCommand now returns its runner who is
called by the main program.

That way the logic to run a SubCommand is moved out of main.rs which
makes for a cleaner design.

For collect subcommand, a little refactor (essentially taking cli out of
Collectors) is needed.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>