module: skb_drop: use a tracepoint instead of a kprobe
Since kfree_skb_reason was introduced in Linux commit c504e5c2f964
("net: skb: introduce kfree_skb_reason()"), the skb:kfree_skb tracepoint
has support for reporting the drop reason. Use this instead of a kprobe
on kfree_skb_reason since this is a more stable interface. The slight
difference is kfree_skb_reason can be called with the SKB_CONSUMED
reason which won't make it into the tp, but that is OK as those are not
for dropped packets.

Signed-off-by: Antoine Tenart <atenart@redhat.com>