core: probe: kernel: run the tracking logic early
Not doing so leads to the tracking logic not running because of
non-fatal errors (e.g. event map is full) which is a big issue as we'll
lose information and some tracking ids won't be freed or handled
correctly.

Fix this by running the tracking logic as soon as possible.

Signed-off-by: Antoine Tenart <atenart@redhat.com>