module: skb_drop: avoid type conversion errors
With recent additions in the kernel, the drop reason can be < 0 (well,
it's a mask, but once converted to an enum value it gets < 0). We still
want for now to keep our key an u32 for easier use later, so just do a
simple fix and exclude potentially problematic values.

We'll improve the logic later, taking into account those masks, but this
is only intended to be a fix.

Signed-off-by: Antoine Tenart <atenart@redhat.com>