module: move factories to Module trait
Right now the modules' section factories are registered independently
and stored in Modules. This doesn't feel all right.
It has to be eventually owned by the BpfEventFactory so we have to
store it as an optional that can be take()en.

In order to cleanup this workaround and make the Modules object more
coherent, add a new function to the Module trait to return the
SectionFactory and create them on the fly.

This refactoring (no behavioral change intended) includes:
- Adding section_factory() to the Module trait
- Renaming as_collector to collector() as it's assuming a Module _is_ a
  Collector and that might not be always the case
- Implement the function for all modules.
- Remove the section_factories from "struct Modules", instead, generate
  them on the fly whenever requested.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>