cli: add helper SubCommand implementation
For simple subcommands, like those that do not need dynamic module
arguments, the current trait forces us to write a lot of boilerplate.

However, clap already transforms a struct into a clap::Command using
derive.

Add a trait that, when implemented by a clap::Parser, it turns it into a
SubCommand that can be used with retis' cli infrastructure.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>