ovs: fix verification failure.
Apparently, a bound check needed by the verifier gets optimized when
compiling with recent clang (Fedora's 16.0.4).
This in turn leads to a load failure because with the following:

85: (71) r2 = *(u8 *)(r3 +15)
invalid access to map value, value_size=331 off=335 size=1
R3 max value is outside of the allowed memory range

Adding a compiler barrier relieves from the issue and makes the
compiler include the instruction needed.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>