core: probe: allow hooks to discard the event
This helps hooks to not only filter themselves but also the full event,
so we don't end up flooded with events in specific cases. This is a
temporary solution and we should aim at at proper one for the longer
term.

Not modify USDT as 1) it's not needed there for now (no users) and 2)
only a single hook per probe is supported anyway.

Signed-off-by: Antoine Tenart <atenart@redhat.com>