process: sort: fix writing result to an non-existing file
In the sort command fs::canonicalize is used to ensure the input and
output files do not match. That work well for existing files but the
mentioned function returns an error in case the file doesn't exist.

In such case we need to construct and clean the path ourselves. We still
rely on fs::canonicalize when possible to handle symlinks.

Signed-off-by: Antoine Tenart <atenart@redhat.com>