usdt: remove ManuallyDrop<> workaround.
Normally when usdts get loaded the usdt manager gets lazily allocated
and subsequently released when the object gets dropped/destroyed.

Before this change, the object got dropped during the .attach()
leading to the release of the underlying usdt manager resource.  The
subsequent drop of the usdt links, which requires the manager in order
to complete all the operations, then led to a dereference of the
dangling pointer previously released and a related segmentation fault.

Fix it by avoiding to release it before it is needed and adding it to
the builder as an Option<>, so that possible subsequent users may
check whether or not it's been released.

Fixes: 74b0dfdcf8f8 ("user: add usdt support")
Fixes: cb6bd2f77ae4 ("usdt: add ManullyDrop to the links vector.")
Signed-off-by: Paolo Valerio <pvalerio@redhat.com>