module: improve can_run checks when kernel config is not available
Do not fail when the kernel config is not available and try to detect if
we can run by looking at a well known symbol. Still keep the kernel
config check if available for better ux. Still, a core level warning is
issued to let the user know some checks won't be made.

This helps running in containers such as toolbox when the kernel config
is not available.

Signed-off-by: Antoine Tenart <atenart@redhat.com>