module: nft: bail out on non-supported older kernels
The nft module do support different kernels types, as long as 'struct
nft_rule_dp' exists. Older kernel have a 'struct nft_rule' instead which
we do not support at the moment. Bail out temporarily until we get
proper support.

This fixes eBPF relocation errors while using the collect command (e.g.
on RHEL9). This did not happen before commit 7d3508f697b0 ("nft: do not
always show the base chain") because another bug made some pointers to
be NULL on such kernels which prevented the probe to go this far.

We bail out instead of fixing the issue because 1) the issue is not
straight forward to fix and 2) the nft module was not working on those
older kernel anyway before and this is only making the error to be
properly supported.

Signed-off-by: Antoine Tenart <atenart@redhat.com>