module: skb_drop: do not fail on old kernels
Do not fail if a kernel is old enough not to support skb drop reasons,
so that the skb-drop reasons still work (installing probes at the right
places) and especially for the dropmon profile to work (which still
provides useful data).

Additionally on kernel not supporting drop reasons, fake reporting the
NOT_SPECIFIED reason.

Signed-off-by: Antoine Tenart <atenart@redhat.com>