module: skb_drop: do not parse the reasons on-demand
Parsing the drop reasons on-demand when the first event arrives can take
quite some time and does not help to dequeue events fast enough. It adds
an uncertainty when starting to collect events and leads to
unpredictable behavior.

Instead, get the drop reasons when initializing the drop reason factory,
only when constructing events from BPF to allow post-processing by
unprivileged users or on different machines.

Signed-off-by: Antoine Tenart <atenart@redhat.com>