core: tracking: skb: special case net:net_dev_xmit
As discussed in an old thread[1] net:net_dev_xmit might access an
already freed skb. Not much we can do here, expect for special casing
the net:net_dev_xmit probe and only report a tracking id if it already
exists. In case the skb was freed by the driver, we can still report an
event if requested, but let's not create a tracking id that won't ever
be freed.

[1] https://lore.kernel.org/lkml/4DE877E1.7000606@jp.fujitsu.com/T/

Signed-off-by: Antoine Tenart <atenart@redhat.com>