collect: cli: set a default value for the out file when requested
This allows to not specify an output file when using the `--out` option.
Defaults to `retis.data`.

Eg,

$ retis collect --out custom.data
$ retis collect -o	# defaults to retis.data
$ retis collect		# no output file

For this to work, both num_args and require_equals needs to be used,
according to the Clap documentation[1]. But using require_equals would
make the --out arg the only one to require using = before the value, and
the use of require_equals with default_missing_value is not actually
enforced by Clap. Moreover not using require_equals is only an issue
when a positional argument is used, which is not the case with collect
(and won't be the case as the default behavior of the command is to do
things automatically).

There is one concern on doing the above: if Clap enforces require_equals
when using default_missing_value in the future. This is fine as we could
implement this manually in collect::cli, by removing
default_missing_value and adding the following:

  if self.args.out.is_none() && args.ids().any(|k| k == "out") {
      self.args.out = Some("retis.data".into());
  }

To be complete and for context, another solution was tested: using an
--out file by default and adding a parameter to disable it (e.g.
--no-out). This doesn't work well with profiles, if a profile wants to
set --no-out while still allowing the user to override that value. This
is because profile arguments are appended to user defined ones, and we
can't do otherwise (at that step /proc/self/cmdline has no meaning, aka
path, command and args are not parsed, and we can't insert the profile
args at the right place).

[1] https://docs.rs/clap/latest/clap/struct.Arg.html#method.default_missing_value

Signed-off-by: Antoine Tenart <atenart@redhat.com>