module: skb_drop: support non-core drop reasons
We could have implemented this using a single HashMap as the (current)
way of defining drop reasons in the Linux kernel makes them
non-overlapping. But Using a per-subsystem representation makes the code
more robust to upstream changes, and to unknown sub-systems.

Added the new "subsys" event field as an Option to keep backward
compatibility and because special casing the core drop reasons seemed a
better representation.

Events now look like:

  22041085978438 [swapper/2] 0 [tp] skb:kfree_skb drop (TCP_INVALID_SEQUENCE)
  22041681941995 [irq/167-iwlwifi] 1242 [tp] skb:kfree_skb drop (mac80211_monitor/RX_DROP_MONITOR)
  22041682024796 [kworker/4:2] 67390 [tp] skb:kfree_skb drop (NOT_SPECIFIED)

Or by faking we do not know about mac80211 reasons,

  WARN  Unknown drop reason subsystem (mac80211_monitor)
  WARN  Unknown drop reason subsystem (mac80211_unusable)

  22946814247525 [irq/167-iwlwifi] 1242 [tp] skb:kfree_skb drop (mac80211_monitor/65538)

Signed-off-by: Antoine Tenart <atenart@redhat.com>