tree: introduce a macro to decorate custom types used in events
Lots of modules & core event sections use custom defined types as
members, all of which have to derive a long list of mandatory traits.
Introduce a macro, event_type, to unify and simplify those custom type
definitions. It might also help in the future if we need to act on the
event definition as a whole.

Signed-off-by: Antoine Tenart <atenart@redhat.com>