cli: keep subcommands ordered
When printing Retis help, subcommands are listed in a random order,
which make the help a bit more difficult to read and impossible to
organize. This is because subcommands are stored in an HashMap.

Converting this to a BTreeMap could work but then subcommands would be
organized by keys, not following a defined order. So instead convert the
subcommand list to be a Vec, which satisfies the above and is a simpler
data struct. The only drawback being checking if a key is already
present, but given the small amount of subcommands and that logic not
running in the hot path this is IMHO perfectly fine.

While at it reorder the profile subcommand.

Signed-off-by: Antoine Tenart <atenart@redhat.com>