tree: do not warn when using the default probe type
When an user does not set a probe type, it defaults to 'kprobe', which
is perfectly fine. In such case an INFO notice is issued.

This bad for UX and using 'kprobe' as the default TYPE really has no
issue, so let's remove those notices and embrace this default value.

Signed-off-by: Antoine Tenart <atenart@redhat.com>