module: skb_drop: do not special case the whole set of drop reasons
Drop reasons are defined in an enum containing u32 values, but we have a
special case where the BPF event contains a negative value (-1). Instead
of special casing the while set of reasons to match this, only special
case the value returned from BPF side and cast it as soon as we can.

Signed-off-by: Antoine Tenart <atenart@redhat.com>