tree: split probe manager into states
The probe manager has two lives, the "builder" state in which probes,
maps & hooks can be added; and the "runtime" one where new elements
can't be added to the manager. The transition between the two states is
when probes, maps and hooks are installed.

This patch makes this more obvious and has multiple advantages, such as
making helpers to be linked to the right probe manager state; avoid any
misused and wrong expectations.

Signed-off-by: Antoine Tenart <atenart@redhat.com>