core: kernel: symbol: require symbols to have a BTF definition
Prevents from ending up with symbols being unusable. On most distros if
a symbol is found in the trace sysfs files, it should have a BTF
definition. However on older kernels, /sys/kernel/btf only has vmlinux
definition and in such cases we wouldn't be able to work with the symbol
later on. Additionally, when /sys/kernel/btf isn't mounted this provides
an additional sanity check.

Signed-off-by: Antoine Tenart <atenart@redhat.com>