module: skb: bpf: fix casting issues
When computing the offset and size of raw packets within the linear part
of an skb, mac and headroom are used. Those arithmetic operations can
issue negative values but variables involved were declared as unsigned
which made some operations to overflow instead.

Fix this by declaring those variables as signed.

Signed-off-by: Antoine Tenart <atenart@redhat.com>