module: skb: always retrieve l3-l4 sections from packets
We used to give a lot of flexibility over which protocols to retrieve
from skbs, mainly as a way to optimize the collection of events as every
protocol was separately handled in BPF. There was no other reason than
that since not displaying those informations is not great for UX.

Since we'll change how we retrieve protocol sections in later patches,
this makes all the protocol sections to always be set and deprecates
the corresponding --skb-sections entries.

For simplicity, the 'eth' section is always retrieved for now but that
will change after the conversion since it is not always useful for the
user.

Signed-off-by: Antoine Tenart <atenart@redhat.com>