Makefile: do not allow to override version
The version should be trusted and its format will be improved in a later
patch. Let's not make overriding it a built-in feature. However as Paolo
noted, let's keep the possibility to override the version name which is
very handy when building custom targets so we can have a human readable
description in addition to the version.

Signed-off-by: Antoine Tenart <atenart@redhat.com>