process: cli: sort: drop path-clean dependency
Path-clean is not equivalent to canonicalize and its `Eq` implementation
can return false for the same file depending on how the path is
expressed therefore its usage was buggy. Drop the path-clean dependency
and rely on fs::canonicalize() instead.

fs::canonicalize() can't be used if the file doesn't exist that is not a
problem for our use case. If the output file doesn't exist it means it
is not the input file, considering we are throwing an error when the
input file doesn't exist.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>