events: display: make the format a struct
DisplayFormat was an enum, which does not modelize well its
capabilities. The format could embed more specific information than a
generic flavor, eg. to control the display of a given type or to give
more control to a sub-section about how the formatting should be done.

This makes it a struct to allow embedding various kinds of information.
The multi vs single line information is now embedded as a boolean.

Signed-off-by: Antoine Tenart <atenart@redhat.com>