Makefile: augment RUSTFLAGS globally in release
Add the warnings flag to RUSTFLAGS globally in the release target to
prevent targets depending on this one to build Retis again when
executed.

For example `make install` was unconditionally rebuilding Retis even if
a release build was done before, this was because the RUSTFLAGS differed
between the targets. We now set it for all future targets, so the build
states are consistent.

Reported-by: Paolo Valerio <pvalerio@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>