tree: add a raw_event_section macro
Use it for all raw event types. Helps in having a unified set of raw
events, all repr as C and all implementing Default. While at it remove
the packed representation for all the event sections that were not yet
converted, except for unions.

Some module use an event sub-section directly. That is fine and
preferable when ever possible.

Signed-off-by: Antoine Tenart <atenart@redhat.com>