retis: benchmark: build the raw event
Move to building the raw event instead of using a static slice or u8.
The raw event does not currently has any section, this will come in next
patches. The main advantage to switch the logic here is maintenance, as
any modification in event sections was impacting the benchmarking code
and updating the static raw event was really difficult (need to generate
one, fixup some parts -such as the symbol addr-, etc).

In addition helpers to facilitate the raw event generation are provided
and a new trait is defined for raw event sections to generate a raw
event.

Signed-off-by: Antoine Tenart <atenart@redhat.com>