Makefile,ebpf.mk: fix/allow cross-compilation
The build system allowed to set the ARCH environment variable, as this
is required for build the BPF programs. But this alone did not allow
cross-compilation. The main issue was setting CARGO_CFG_TARGET_ARCH has
no effect and should be used the other way around: it is exported by
cargo to allow build script to know the target arch.

In order to allow cross-compilation, we should let Cargo/Rustc know what
should be the target. The target not only has information about the
architecture but also contains information about the target system and
abi. It can be specified as --target or using the CARGO_BUILD_TARGET
environment variable.

As Retis is a Rust application and as the target triple contains all the
information we need to correctly cross compile an application without
making (wrong) assumption, this patch chooses to use CARGO_BUILD_TARGET
as our top level cross-compilation configuration. The ARCH variable
(used by the BPF build part) is now derived from it. If
CARGO_BUILD_TARGET is not provided, we keep the same logic and use the
default target of the current Rust compiler.

Signed-off-by: Antoine Tenart <atenart@redhat.com>