events: kernel: add common types factory
As explained in the code,

Logic to retrieve common types from our probes whenever such type is seen
(initial use case is as function parameter, but this could be extended to
struct members). This is intended for small types (eg. enums) that are not
found all over the place and will always very likely provide useful
information. They have to fit in an u64 anyway.

Signed-off-by: Antoine Tenart <atenart@redhat.com>