probe: kernel: move drop reason logic to common types
This makes the skb-drop module to only install a probe on the
skb:kfree_skb tracepoint from now on. The logic to parse the drop reason
is moved to the core and to the new common types infrastructure. This
makes drop reasons to be always retrieved whenever they are available.

In addition to the above, no warning is issued if drop reasons are not
available on a specific kernel. This is better than generating fake
events and maps better to the actual kernel features.

Strictly speaking having a module to only install a probe on the above
tracepoint is a bit complex, but that is just one additional example
showing the limits of our module abstraction.

Signed-off-by: Antoine Tenart <atenart@redhat.com>