events: continue to special case the drop reason formatting
One of the issues with the current logic is section ids are linked to
both the BPF event factories and the event sections themselves. This
prevents many things, including a factory generating more than a single
section (and using a different section id).

Partially workaround this for now by s/SkbDrop/CommonType/ in the
formatting logic and by making SectionId::CommonType a special value
(same as other ones, like SectionId::SkbDrop).

The side effect is to make reset reasons to behave like the drop
reasons, which IMO is fine.

Signed-off-by: Antoine Tenart <atenart@redhat.com>