tree: link event sections to section ids
Those two were loosely tightened, when adding sub-sections to an event.
Even more, the section names where both defined in the event sections
and in the section ids.

As there should be a 1:1 relationship between event sections and ids,
make the id part of the event section.

This has another advantage as a side effect: raw section factories do
not need a 1:1 relationship with an event section and can generate any
event section.

Signed-off-by: Antoine Tenart <atenart@redhat.com>