tree: untangle factories from event sections
Stop using SectionId for indexing event section factories; those are
tied to a BPF counterpart, not to a section id. This helps modelizing
that a factory can produce different sections and that a section does
not necessarily have a BPF counterpart, such as the Tracking one.

To do this use a dedicated FactoryId enum instead of reusing SectionId
and use a similar logic as what we do for event sections (eg. embed the
id into the factory definition).

While doing this we can remove the TrackingInfoEventFactory.

Signed-off-by: Antoine Tenart <atenart@redhat.com>