python: add python event representation.
Use pyo3 to represent events and event sections.

Notably, this requires some simple enums to implement PartialEq.

EventTypes are given a __repr__() helper that uses the debug formatter.
It seems a good first start.

EventSections are given a __repr__() helper that prints a dictionary
based on the json object as well as a show() helper that formats the
event.

Co-authored-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>
Signed-off-by: Adrian Moreno <amorenoz@redhat.com>