retis: module: skb: if the packet is not supported at least provide the etype
$subject ; this fixes an issue where if a packet was not parsed by the
skb module, the output would contain blank lines. This was not
convenient and more importantly hides a real issue: the user had no clue
what this packet was.

Fix this by parsing the ethernet header (might be fake, but with at
least a real etype) in such case, even if it was not requested.

Signed-off-by: Antoine Tenart <atenart@redhat.com>