retis: module: skb: if the packet is not supported at least provide the etype
If we get a packet we don't recognize and don't know how to parse, the
output of the skb section would be empty and we'll end up with extra
blank lines. This is not convenient but more importantly hides a real
issue: the user had no clue what this packet is.

Fix this by parsing the ethernet header (might be fake, but could have a
real etype) in such case, even if it was not requested. Only report
results if the ethertype looks valid, otherwise it might not be an
ethernet packet (or one with invalid offsets). It could still be invalid
(only looking like a valid ethertype) but what can we do?

Signed-off-by: Antoine Tenart <atenart@redhat.com>