retis: module: skb: do not report ethernet packet if they do not look like one
In the previous commit we added reporting the ethernet header if it
looked like one, to improve user experience. While looking at it I
realized if `--skb-section eth` was used we could also report invalid
ethernet data.

Widen the sanity check for ethernet headers to also take the
`--skb-section eth` into account. This has the advantage of being
consistent and provides a good argument for trying to report the
ethernet header if we couldn't parse the packet (because doing so could
report an invalid one in rare cases, but so the `--skb-section eth`
case).

Signed-off-by: Antoine Tenart <atenart@redhat.com>