tools: localversion: handle shallow clones
When a shallow clone is done (--depth X) there might not be any tag
reported in the history. The same can happen when cloning w/o tags
(--no-tags). In such case `git describe` will print an error.

While this does not affect the result of the localversion script, it
makes the output to be filled with error messages. Fix this by ignoring
error messages from that command.

Signed-off-by: Antoine Tenart <atenart@redhat.com>