workflow: split image build into two workflows
The first one is generic and checks for new commits in main since the
day before. If there are, it triggers the building image one.

This helps not having failed or canceled workflows (thus notifications
and also proper monitoring of the workflows) when no new commit is
found. Also this opens the way for other workflows if needed.

Signed-off-by: Antoine Tenart <atenart@redhat.com>