retis: auto-detect user provided tracepoints
This makes the probe type optional when using a tracepoint; one can use
`-p skb:kfree_skb` instead of `-p tp:skb:kfree_skb`.

The logic is as follow:
- If a probe type is explicitly given, we're using it.
- If no probe type is given:
  + If a single ':' is found, defaults to 'raw_tracepoint'.
  + If no ':' is found, defaults to 'kprobe'.

Note that default probe types are not considered stable and might evolve
over time (and this was never the intent). Update the help accordingly.

Signed-off-by: Antoine Tenart <atenart@redhat.com>